+++ HISTORY.rst
@@ -4,6 +4,7 @@ Release history
 Current release
 ---------------
 
+* compat module is deprecated for 5 years and will be removed with next release (T183085)
 * ipaddress module is required for Python 2 (T243171)
 * tools.ip will be dropped in favour of tools.is_IP (T243171)
 * tools.ip_regexp is deprecatd for 5 years and will be removed with next release